Mouse wheel in WASM. Also changed mouse wheel to int and made it give consistent values between Windows and WASM: generally 1 or -1.

+++ app/app.odin
@@ -548,7 +548,7 @@ middle_mouse_double_click :: proc "contextless" () -> bool {
 }
 
 // Returns mouse wheel delta as a normalized floating point value.
-mouse_wheel :: proc "contextless" () -> f32 {
+mouse_wheel :: proc "contextless" () -> int {
     return ctx.mouse_wheel
 }
